Restructure. Configuring test environment.
