+++ app/screens/home/Home.spec.tsx
@@ -1,0 +1,8 @@
+import Home from './Home';
+import { render } from '@testing-library/react';
+
+describe('<Home />', () => {
+  it('renders the component', () => {
+    render(<Home />);
+  });
+});
