+++ package.json
@@ -7,15 +7,17 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "lint": "eslint \"**/*.{js,jsx,ts,tsx}\"",
-    "lint:fix": "eslint --fix \"**/*.{js,jsx,ts,tsx}\"",
     "format": "npx prettier --check \"**/*.{js,jsx,ts,tsx}\"",
     "format:fix": "npx prettier --write \"**/*.{js,jsx,ts,tsx}\"",
-    "clean": "npx prettier --write \"**/*.{js,jsx,ts,tsx}\" && npm run lint:fix"
+    "lint": "eslint \"**/*.{js,jsx,ts,tsx}\"",
+    "lint:fix": "eslint --fix \"**/*.{js,jsx,ts,tsx}\"",
+    "clean": "npx prettier --write \"**/*.{js,jsx,ts,tsx}\" && npm run lint:fix",
+    "test": "jest"
   },
   "dependencies": {
     "expo": "~48.0.9",
     "expo-status-bar": "~1.4.4",
+    "jest-expo": "^48.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.71.4",
@@ -23,17 +25,17 @@
   },
   "devDependencies": {
     "@babel/core": "^7.21.3",
+    "@testing-library/react": "^14.0.0",
     "@types/jest": "^29.5.0",
-    "@types/react": "~18.0.28",
     "@types/react-native": "~0.71.4",
+    "@types/react": "~18.0.28",
     "@typescript-eslint/eslint-plugin": "^5.56.0",
     "@wildpastry/eslint-config": "^1.6.4",
-    "eslint": "^8.36.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
+    "eslint": "^8.36.0",
     "jest": "^29.5.0",
     "prettier": "^2.8.6",
-    "ts-jest": "^29.0.5",
     "typescript": "^4.9.4"
   },
   "private": true
